Everything is working now.  Navigation is great.  Everything loads. so, what is next?
I need to make sure that it is displaying everything in the correct order.  Currently, on the item display, it shows the different items in a different order than they were added.  I think this is to do with the getKeys(), i think that is what it is called, and this method returns all the keys to the hashSet where all the data is displayed.

I need to also stack a few more lists to ensure that everything is working as it should be with navigation.